Decouple @waveform-playlist/spectrogram from @waveform-playlist/browser
Make spectrogram a truly optional package by introducing a
SpectrogramIntegration React context. The browser package defines the
context interface; the spectrogram package provides SpectrogramProvider
that fills it. When no provider is present, all spectrogram features
are cleanly skipped.

- Create SpectrogramIntegrationContext in browser package
- Move ~600 lines of spectrogram logic into SpectrogramProvider
- Remove @waveform-playlist/spectrogram from browser dependencies
- Update PlaylistVisualization to read from integration context
- Update MIR example to use SpectrogramProvider wrapper
