+++ packages/browser/src/index.tsx
@@ -135,9 +135,9 @@ export type { RenderAnnotationItemProps, AnnotationData } from '@waveform-playli
 // Export annotation callback types for Waveform components
 export type { GetAnnotationBoxLabelFn, OnAnnotationUpdateFn } from './types/annotations';
 
-// Re-export spectrogram computation and utilities
-export { computeSpectrogram, computeSpectrogramMono, getColorMap, getFrequencyScale } from '@waveform-playlist/spectrogram';
-export type { FrequencyScaleName } from '@waveform-playlist/spectrogram';
+// Spectrogram integration context (for optional spectrogram support)
+export { SpectrogramIntegrationProvider, useSpectrogramIntegration } from './SpectrogramIntegrationContext';
+export type { SpectrogramIntegration } from './SpectrogramIntegrationContext';
 
 // Export waveform-data.js utilities
 export {
